+++ react/components/OrganizationUsersTable.tsx
@@ -109,7 +109,10 @@ const OrganizationUsersTable: FunctionComponent<Props> = ({
     }
   }
 
+  const canManageOrganization = permissions.includes('manage-organization')
+
   const canEdit = isAdmin || permissions.includes('add-users-organization')
+
   const canEditSales =
     isAdmin ||
     permissions.includes('add-sales-users-all') ||
@@ -394,14 +397,19 @@ const OrganizationUsersTable: FunctionComponent<Props> = ({
   }
 
   const getSchema = () => {
-    const isEnabled = ({ role: { slug } }: { role: { slug: string } }) =>
-      ruleClickEnabled({
-        isAdmin,
-        canEditSales,
-        slug,
-        canEdit,
-        isSalesAdmin,
-      })
+    const isEnabled = ({ role: { slug } }: { role: { slug: string } }) => {
+      if (!canManageOrganization) {
+        return ruleClickEnabled({
+          isAdmin,
+          canEditSales,
+          slug,
+          canEdit,
+          isSalesAdmin,
+        })
+      }
+
+      return canManageOrganization
+    }
 
     const properties = {
       email: {
@@ -542,7 +550,20 @@ const OrganizationUsersTable: FunctionComponent<Props> = ({
   }
 
   const handleRowClick = ({ rowData }: CellRendererProps) => {
-    if (
+    if (canManageOrganization) {
+      setEditUserDetails({
+        id: rowData.id,
+        roleId: rowData.roleId,
+        userId: rowData.userId,
+        clId: rowData.clId,
+        orgId: rowData.orgId,
+        costId: rowData.costId,
+        name: rowData.name,
+        email: rowData.email,
+        canImpersonate: rowData.canImpersonate,
+      })
+      setEditUserModalOpen(true)
+    } else if (
       !ruleClickEnabled({
         isAdmin,
         canEditSales,
@@ -551,21 +572,9 @@ const OrganizationUsersTable: FunctionComponent<Props> = ({
         isSalesAdmin,
       })
     ) {
+      // eslint-disable-next-line no-useless-return
       return
     }
-
-    setEditUserDetails({
-      id: rowData.id,
-      roleId: rowData.roleId,
-      userId: rowData.userId,
-      clId: rowData.clId,
-      orgId: rowData.orgId,
-      costId: rowData.costId,
-      name: rowData.name,
-      email: rowData.email,
-      canImpersonate: rowData.canImpersonate,
-    })
-    setEditUserModalOpen(true)
   }
 
   const handleSort = ({
@@ -665,6 +674,7 @@ const OrganizationUsersTable: FunctionComponent<Props> = ({
           canEdit={canEdit}
           canEditSales={canEditSales}
           isSalesAdmin={isSalesAdmin}
+          canManageOrg={canManageOrganization}
         />
       )}
       {editUserModalOpen && (
@@ -680,6 +690,7 @@ const OrganizationUsersTable: FunctionComponent<Props> = ({
           canEdit={canEdit}
           canEditSales={canEditSales}
           isSalesAdmin={isSalesAdmin}
+          canManageOrg={canManageOrganization}
         />
       )}
       <RemoveUserModal